+++ plugin-dev/Source/Sentry/Private/Android/Infrastructure/SentryJavaClasses.cpp
@@ -14,6 +14,7 @@ const FSentryJavaClass SentryJavaClasses::User				= FSentryJavaClass { "io/sentr
 const FSentryJavaClass SentryJavaClasses::UserFeedback		= FSentryJavaClass { "io/sentry/UserFeedback", ESentryJavaClassType::External };
 const FSentryJavaClass SentryJavaClasses::Message			= FSentryJavaClass { "io/sentry/protocol/Message", ESentryJavaClassType::External };
 const FSentryJavaClass SentryJavaClasses::SentryLevel		= FSentryJavaClass { "io/sentry/SentryLevel", ESentryJavaClassType::External };
+const FSentryJavaClass SentryJavaClasses::SentryHint		= FSentryJavaClass { "io/sentry/Hint", ESentryJavaClassType::External };
 
 // System Java classes definitions
 const FSentryJavaClass SentryJavaClasses::ArrayList			= FSentryJavaClass { "java/util/ArrayList", ESentryJavaClassType::System };