+++ plugin-dev/Source/Sentry/Private/Desktop/SentryEventDesktop.h
@@ -21,6 +21,7 @@ public:
 	virtual FString GetMessage() const override;
 	virtual void SetLevel(ESentryLevel level) override;
 	virtual ESentryLevel GetLevel() const override;
+	virtual bool IsCrash() const override;
 
 private:
 	sentry_value_t EventDesktop;