+++ plugin-dev/Source/Sentry/Private/Desktop/SentrySubsystemDesktop.h
@@ -13,7 +13,7 @@ class SentrySubsystemDesktop : public ISentrySubsystem
 public:
 	SentrySubsystemDesktop();
 
-	virtual void InitWithSettings(const USentrySettings* settings) override;
+	virtual void InitWithSettings(const USentrySettings* settings, USentryBeforeSendHandler* beforeSendHandler) override;
 	virtual void Close() override;
 	virtual bool IsEnabled() override;
 	virtual void AddBreadcrumb(USentryBreadcrumb* breadcrumb) override;