+++ plugin-dev/Source/Sentry/Private/Interface/SentryHintInterface.h
@@ -1,0 +1,15 @@
+// Copyright (c) 2023 Sentry. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+
+class USentryAttachment;
+
+class ISentryHint
+{
+public:
+	virtual ~ISentryHint() = default;
+
+	virtual void AddAttachment(USentryAttachment* attachment) = 0;
+};